requirements.txt
confusable-homoglyphs==3.2.0
Django==1.11.5
django-bootstrap3==11.0.0
django-material==1.5.2
django-registration==2.4.1
django-smartfields==1.0.9
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
whitenoise==3.3.1
